Port `do` snippet
